+++ core/src/main/java/app/cash/paykit/core/models/request/CustomerRequestDataFactory.kt
@@ -37,9 +37,16 @@ object CustomerRequestDataFactory {
     isRequestUpdate: Boolean = false,
   ): CustomerRequestData {
     val actions = ArrayList<Action>(paymentActions.size)
+    var possibleReferenceId: String? = null
+
     for (paymentAction in paymentActions) {
       when (paymentAction) {
-        is OnFileAction -> actions.add(buildFromOnFileAction(clientId, paymentAction))
+        is OnFileAction -> {
+          actions.add(buildFromOnFileAction(clientId, paymentAction))
+          if (paymentAction.accountReferenceId != null) {
+            possibleReferenceId = paymentAction.accountReferenceId
+          }
+        }
         is OneTimeAction -> actions.add(buildFromOneTimeAction(clientId, paymentAction))
       }
     }
@@ -49,12 +56,14 @@ object CustomerRequestDataFactory {
         actions = actions,
         channel = null,
         redirectUri = null,
+        referenceId = possibleReferenceId?.let { PiiString(it) },
       )
     } else {
       CustomerRequestData(
         actions = actions,
         channel = CHANNEL_IN_APP,
-        redirectUri = redirectUri?.let { PiiString(redirectUri) },
+        redirectUri = redirectUri?.let { PiiString(it) },
+        referenceId = possibleReferenceId?.let { PiiString(it) },
       )
     }
   }